Refactor: Move RentalControl to a coordinator
Move the RentalControl class to a new file to make it easier to manage
and maintain. This will also allow for the class to be used in multiple
places in the future.

Cleanup latest mypy errors that would stop the code from passing
linting.
